feat: add dockerfile to the templates
