+++ _templates/backend/basic/config/config.json.ejs.t
