+++ _templates/backend/module/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+    params: ({ args }) => {
+        const parsedConfig = JSON.parse(args.config);
+        console.log(parsedConfig);
+        return parsedConfig;
+    }
+};
