+++ _templates/backend/module/src/models/color.ejs.t
