--- _templates/database/many/index.js
@@ -1,7 +1,0 @@
-module.exports = {
-  params: ({ args }) => {
-    const parsedConfig = JSON.parse(args.config);
-    console.log('parsedConfig', parsedConfig.database);
-    return parsedConfig.database;
-  }    
-};
