+++ _templates/database/module/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+    params: ({ args }) => {
+        const parsedConfig = JSON.parse(args.config);
+        console.log('parsedConfig', parsedConfig.database);
+        return parsedConfig.database;
+    }    
+};
