--- _templates/database/new/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  params: ({ args }) => {
-    const parsedConfig = JSON.parse(args.config);
-    return parsedConfig.database;
-  }    
-};
