+++ _templates/frontend/basic/app/package.json.ejs.t
