+++ _templates/frontend/basic/app/src/App.ejs.t
