+++ _templates/frontend/basic/app/src/components/footer.ejs.t
