+++ _templates/frontend/basic/app/src/components/header.ejs.t
