+++ _templates/frontend/basic/app/src/components/healthCheck.ejs.t
