+++ _templates/frontend/basic/app/src/components/main.ejs.t
