+++ _templates/frontend/basic/app/src/constants/apiEndPoints.ejs.t
