+++ _templates/frontend/basic/app/src/index.js.ejs.t
