+++ _templates/frontend/basic/app/src/pages/ping.ejs.t
