+++ _templates/frontend/basic/docker.ejs.t
@@ -1,0 +1,17 @@
+---
+to: <%= outputPath %>/<%= appName %>/Dockerfile
+force: true
+---
+FROM alpine:latest
+
+RUN apk add --update nodejs npm
+
+WORKDIR /app
+
+COPY package.json package.json
+
+RUN npm install
+
+COPY . .
+
+ENTRYPOINT ["npm", "start"]
