+++ _templates/frontend/basic/index.js
