+++ tests/services/generators/k8s-manifest.test.js
@@ -1,21 +1,24 @@
-const k8sManifestGenerator = require("../../../src/services/generators/k8s-manifest");
+const { spawn } = require('child_process');
+const { stderr } = require('process');
+
+const k8sManifestGenerator = require('../../../src/services/generators/k8s-manifest');
 
 let mockSpawn = {};
 let mockWriteStream = {};
 
-jest.mock("fs", () => ({
+jest.mock('fs', () => ({
   createWriteStream: () => mockWriteStream,
 }));
 
-jest.mock("child_process", () => ({
+jest.mock('child_process', () => ({
   spawn: () => mockSpawn,
 }));
 
-describe("k8sManifestGenerator", () => {
-  it("should generate a k8s-manifest.yaml file", async () => {
+describe('k8sManifestGenerator', () => {
+  it('should generate a k8s-manifest.yaml file', async () => {
     mockSpawn = {
       on: jest.fn().mockImplementation((event, callback) => {
-        if (event === "close") {
+        if (event === 'close') {
           callback(0);
         }
       }),
@@ -28,10 +31,10 @@ describe("k8sManifestGenerator", () => {
     expect(code).toBe(0);
   });
 
-  it("should throw error when the k8s-manifest is not generated", async () => {
+  it('should throw error when the k8s-manifest is not generated', async () => {
     mockSpawn = {
       on: jest.fn().mockImplementation((event, callback) => {
-        if (event === "close") {
+        if (event === 'close') {
           callback(1);
         }
       }),
@@ -47,16 +50,16 @@ describe("k8sManifestGenerator", () => {
     }
   });
 
-  it("should throw an error if docker-compose.yaml file does not exist", async () => {
+  it('should throw an error if docker-compose.yaml file does not exist', async () => {
     mockSpawn = {
       on: jest.fn().mockImplementation((event, callback) => {
-        if (event === "close") {
+        if (event === 'close') {
           callback(1);
         }
       }),
       stdout: {
         pipe: jest.fn().mockImplementation(() => {
-          throw new Error("error");
+          throw new Error('error');
         }),
       },
     };
@@ -64,7 +67,7 @@ describe("k8sManifestGenerator", () => {
     try {
       await k8sManifestGenerator(1);
     } catch (e) {
-      expect(e.message).toBe("error");
+      expect(e.message).toBe('error');
     }
   });
 });
